#76 refactor: 연도 헤더 컴포넌트 분리
