+++ src/page/archiving/component/YearHeader/YearHeader.style.ts
@@ -1,0 +1,7 @@
+import { css } from '@emotion/react';
+
+export const headerStyle = css({
+  display: 'flex',
+
+  padding: '0rem 0rem 2.4rem 0rem',
+});
